+++ stars-logic-kcmftbl/src/main/kotlin/tools/aqua/stars/logic/kcmftbl/pastMinPrevalence.kt
@@ -15,13 +15,15 @@
  * limitations under the License.
  */
 
+@file:Suppress("unused")
+
 package tools.aqua.stars.logic.kcmftbl
 
 import tools.aqua.stars.core.types.*
 
 /**
- * CMFTBL implementation of the 'once' operator i.e. "In a past tick in the interval phi holds at
- * least once".
+ * CMFTBL implementation of the 'pastMinPrevalence' operator i.e. "In all past ticks in the interval
+ * phi holds for at least ([percentage]*100)% of the ticks in the interval".
  *
  * @param E [EntityType].
  * @param T [TickDataType].
@@ -29,6 +31,7 @@ import tools.aqua.stars.core.types.*
  * @param U [TickUnit].
  * @param D [TickDifference].
  * @param tickData Current [TickDataType].
+ * @param percentage Threshold value.
  * @param interval Observation interval.
  * @param phi Predicate.
  */
@@ -37,16 +40,39 @@ fun <
     T : TickDataType<E, T, S, U, D>,
     S : SegmentType<E, T, S, U, D>,
     U : TickUnit<U, D>,
-    D : TickDifference<D>> once(
+    D : TickDifference<D>> pastMinPrevalence(
     tickData: T,
+    percentage: Double,
     interval: Pair<D, D>? = null,
     phi: (T) -> Boolean
-): Boolean =
-    since(tickData = tickData, interval = interval, phi1 = { _ -> true }, phi2 = { td -> phi(td) })
+): Boolean {
+  checkInterval(interval)
+  checkPercentage(percentage)
+
+  val segment = tickData.segment
+  val now = tickData.currentTick
+  val nowIndex = segment.tickData.indexOf(tickData)
+
+  var tickCount = 0
+  val trueCount =
+      (0..nowIndex).count { currentIndex ->
+        val currentTickData = segment.tickData[currentIndex]
+
+        if (interval != null &&
+            (currentTickData.currentTick > now - interval.first ||
+                currentTickData.currentTick <= now - interval.second))
+            return@count false
+
+        tickCount++
+        phi(currentTickData)
+      }
+
+  return trueCount >= tickCount * percentage
+}
 
 /**
- * CMFTBL implementation of the 'once' operator for one entity i.e. "In a past tick in the interval
- * phi holds at least once".
+ * CMFTBL implementation of the 'minPrevalence' operator for one entity i.e. "In all past ticks in
+ * the interval phi holds for at least ([percentage]*100)% of the ticks in the interval".
  *
  * @param E1 [EntityType].
  * @param E [EntityType].
@@ -55,25 +81,32 @@ fun <
  * @param U [TickUnit].
  * @param D [TickDifference].
  * @param entity Current [EntityType] of which the tickData gets retrieved.
+ * @param percentage Threshold value.
  * @param interval Observation interval.
  * @param phi Predicate.
  */
+@Suppress("UNCHECKED_CAST")
 fun <
     E1 : E,
     E : EntityType<E, T, S, U, D>,
     T : TickDataType<E, T, S, U, D>,
     S : SegmentType<E, T, S, U, D>,
     U : TickUnit<U, D>,
-    D : TickDifference<D>> once(
+    D : TickDifference<D>> pastMinPrevalence(
     entity: E1,
+    percentage: Double,
     interval: Pair<D, D>? = null,
     phi: (E1) -> Boolean
 ): Boolean =
-    since(entity = entity, interval = interval, phi1 = { _ -> true }, phi2 = { e -> phi(e) })
+    pastMinPrevalence(
+        tickData = entity.tickData,
+        percentage = percentage,
+        interval = interval,
+        phi = { td -> td.getEntityById(entity.id)?.let { phi(it as E1) } == true })
 
 /**
- * CMFTBL implementation of the 'once' operator for two entities i.e. "In a past tick in the
- * interval phi holds at least once".
+ * CMFTBL implementation of the 'minPrevalence' operator for two entities i.e. "In all past ticks in
+ * the interval phi holds for at least ([percentage]*100)% of the ticks in the interval".
  *
  * @param E1 [EntityType].
  * @param E2 [EntityType].
@@ -84,9 +117,11 @@ fun <
  * @param D [TickDifference].
  * @param entity1 First [EntityType].
  * @param entity2 Second [EntityType].
+ * @param percentage Threshold value.
  * @param interval Observation interval.
  * @param phi Predicate.
  */
+@Suppress("UNCHECKED_CAST", "DuplicatedCode")
 fun <
     E1 : E,
     E2 : E,
@@ -94,15 +129,24 @@ fun <
     T : TickDataType<E, T, S, U, D>,
     S : SegmentType<E, T, S, U, D>,
     U : TickUnit<U, D>,
-    D : TickDifference<D>> once(
+    D : TickDifference<D>> pastMinPrevalence(
     entity1: E1,
     entity2: E2,
+    percentage: Double,
     interval: Pair<D, D>? = null,
     phi: (E1, E2) -> Boolean
-): Boolean =
-    since(
-        entity1 = entity1,
-        entity2 = entity2,
-        interval = interval,
-        phi1 = { _, _ -> true },
-        phi2 = { e1, e2 -> phi(e1, e2) })
+): Boolean {
+  require(entity1.tickData == entity2.tickData) {
+    "the two entities provided as argument are not from same tick"
+  }
+  return pastMinPrevalence(
+      tickData = entity1.tickData,
+      percentage = percentage,
+      interval = interval,
+      phi = { td ->
+        val pastEntity1 = td.getEntityById(entity1.id)
+        val pastEntity2 = td.getEntityById(entity2.id)
+        if (pastEntity1 == null || pastEntity2 == null) false
+        else phi(pastEntity1 as E1, pastEntity2 as E2)
+      })
+}